+++ workspace/widgets/metadata/src/utils/export.ts
@@ -1,3 +1,20 @@
+/**
+ * Copyright (C) 2023 Zuoqiu Yingyi
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import yaml from "js-yaml";
 import moment from "moment";
 
@@ -7,8 +24,8 @@ import {
     looseJsonParse,
 } from "./string";
 
-import { IForm, IAttr } from "./../types/form";
-import { IMetadata } from "./../types/metadata";
+import { IForm, IAttr } from "@/types/form";
+import { IMetadata } from "@/types/metadata";
 
 /* 解析器 */
 export enum Parser {